+++ airbyte-container-orchestrator/src/main/java/io/airbyte/container_orchestrator/DbtJobOrchestrator.java
@@ -8,8 +8,8 @@ import io.airbyte.config.Configs;
 import io.airbyte.config.OperatorDbtInput;
 import io.airbyte.scheduler.models.IntegrationLauncherConfig;
 import io.airbyte.scheduler.models.JobRunConfig;
-import io.airbyte.workers.DbtTransformationRunner;
-import io.airbyte.workers.DbtTransformationWorker;
+import io.airbyte.workers.general_workers.DbtTransformationRunner;
+import io.airbyte.workers.general_workers.DbtTransformationWorker;
 import io.airbyte.workers.WorkerConfigs;
 import io.airbyte.workers.WorkerUtils;
 import io.airbyte.workers.normalization.NormalizationRunnerFactory;
