+++ airbyte-container-orchestrator/src/main/java/io/airbyte/container_orchestrator/DefaultAsyncStateManager.java
@@ -6,7 +6,7 @@ package io.airbyte.container_orchestrator;
 
 import io.airbyte.workers.process.AsyncKubePodStatus;
 import io.airbyte.workers.process.KubePodInfo;
-import io.airbyte.workers.storage.DocumentStoreClient;
+import io.airbyte.workers.DocumentStoreClient;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
