+++ airbyte-container-orchestrator/src/main/java/io/airbyte/container_orchestrator/ReplicationJobOrchestrator.java
@@ -10,7 +10,7 @@ import io.airbyte.config.ReplicationOutput;
 import io.airbyte.config.StandardSyncInput;
 import io.airbyte.scheduler.models.IntegrationLauncherConfig;
 import io.airbyte.scheduler.models.JobRunConfig;
-import io.airbyte.workers.DefaultReplicationWorker;
+import io.airbyte.workers.general_workers.DefaultReplicationWorker;
 import io.airbyte.workers.RecordSchemaValidator;
 import io.airbyte.workers.ReplicationWorker;
 import io.airbyte.workers.WorkerConfigs;
