+++ airbyte-integrations/bases/standard-destination-test/src/main/java/io/airbyte/integrations/standardtest/destination/DestinationAcceptanceTest.java
@@ -43,9 +43,9 @@ import io.airbyte.protocol.models.DestinationSyncMode;
 import io.airbyte.protocol.models.Field;
 import io.airbyte.protocol.models.JsonSchemaType;
 import io.airbyte.protocol.models.SyncMode;
-import io.airbyte.workers.DbtTransformationRunner;
-import io.airbyte.workers.DefaultCheckConnectionWorker;
-import io.airbyte.workers.DefaultGetSpecWorker;
+import io.airbyte.workers.general_workers.DbtTransformationRunner;
+import io.airbyte.workers.general_workers.DefaultCheckConnectionWorker;
+import io.airbyte.workers.general_workers.DefaultGetSpecWorker;
 import io.airbyte.workers.WorkerConfigs;
 import io.airbyte.workers.WorkerException;
 import io.airbyte.workers.normalization.NormalizationRunner;
@@ -55,7 +55,7 @@ import io.airbyte.workers.process.DockerProcessFactory;
 import io.airbyte.workers.process.ProcessFactory;
 import io.airbyte.workers.protocols.airbyte.AirbyteDestination;
 import io.airbyte.workers.protocols.airbyte.DefaultAirbyteDestination;
-import io.airbyte.workers.test_helpers.EntrypointEnvChecker;
+import io.airbyte.workers.helper.EntrypointEnvChecker;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
