+++ airbyte-workers/src/main/java/io/airbyte/workers/DocumentStoreClient.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers.storage;
+package io.airbyte.workers;
 
 import java.util.Optional;
