+++ airbyte-workers/src/main/java/io/airbyte/workers/general_workers/EchoWorker.java
@@ -2,9 +2,11 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers;
+package io.airbyte.workers.general_workers;
 
 import java.nio.file.Path;
+
+import io.airbyte.workers.Worker;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
