+++ airbyte-workers/src/main/java/io/airbyte/workers/helper/EntrypointEnvChecker.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers.test_helpers;
+package io.airbyte.workers.helper;
 
 import io.airbyte.workers.WorkerException;
 import io.airbyte.workers.process.ProcessFactory;
