+++ airbyte-workers/src/main/java/io/airbyte/workers/process/AsyncOrchestratorPodProcess.java
@@ -9,7 +9,7 @@ import io.airbyte.commons.json.Jsons;
 import io.airbyte.config.ResourceRequirements;
 import io.airbyte.config.helpers.LogClientSingleton;
 import io.airbyte.workers.WorkerApp;
-import io.airbyte.workers.storage.DocumentStoreClient;
+import io.airbyte.workers.DocumentStoreClient;
 import io.fabric8.kubernetes.api.model.ContainerBuilder;
 import io.fabric8.kubernetes.api.model.ContainerPort;
 import io.fabric8.kubernetes.api.model.DeletionPropagation;
