+++ airbyte-workers/src/main/java/io/airbyte/workers/storage/DockerComposeDocumentStoreClient.java
@@ -5,6 +5,8 @@
 package io.airbyte.workers.storage;
 
 import io.airbyte.commons.io.IOs;
+import io.airbyte.workers.DocumentStoreClient;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
