+++ airbyte-workers/src/main/java/io/airbyte/workers/storage/GcsDocumentStoreClient.java
@@ -10,6 +10,8 @@ import com.google.cloud.storage.BlobInfo;
 import com.google.cloud.storage.Storage;
 import io.airbyte.config.storage.CloudStorageConfigs.GcsConfig;
 import io.airbyte.config.storage.DefaultGcsClientFactory;
+import io.airbyte.workers.DocumentStoreClient;
+
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.util.Optional;
